requirements.txt
# Core Framework
fastapi>=0.116.0,<0.117.0
uvicorn[standard]>=0.35.0,<0.36.0
pydantic>=2.11.7,<3.0.0
pydantic-settings>=2.10.1,<3.0.0

# Database
sqlalchemy>=2.0.41,<3.0.0
alembic>=1.16.2,<2.0.0
asyncpg>=0.30.0,<0.31.0
psycopg2-binary>=2.9.10,<3.0.0
pgvector>=0.2.4,<0.3.0

# Vector Databases
qdrant-client>=1.14.3,<2.0.0
pinecone-client>=6.0.0,<7.0.0
chromadb>=1.0.15,<2.0.0

# Cache & Message Queue
redis[hiredis]>=6.2.0,<7.0.0
celery>=5.5.3,<6.0.0
flower>=2.0.1,<3.0.0

# Storage
boto3>=1.35.127,<2.0.0
aiofiles>=24.1.0,<25.0.0

# Security
python-jose[cryptography]>=3.4.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.18,<0.1.0

# AI/ML
openai>=1.93.1,<2.0.0
anthropic>=0.57.1,<1.0.0
tiktoken>=0.8.1,<0.9.0
langchain>=0.3.26,<0.4.0
langchain-community>=0.3.14,<0.4.0
numpy>=1.26.4,<2.0.0
pandas>=2.2.4,<3.0.0
scikit-learn>=1.6.1,<2.0.0

# HTTP Client
httpx>=0.28.1,<0.29.0

# Monitoring & Logging
prometheus-client>=0.21.2,<0.22.0
prometheus-fastapi-instrumentator>=7.0.0,<8.0.0
sentry-sdk>=2.19.1,<3.0.0
structlog>=25.4.0,<26.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
pyyaml>=6.0.2,<7.0.0
click>=8.1.8,<9.0.0

# API Documentation
python-multipart>=0.0.18,<0.1.0

# Additional Production Dependencies
tenacity>=9.1.2,<10.0.0  # Retry library
pydantic-extra-types>=2.12.0,<3.0.0  # Extra Pydantic types
email-validator>=2.2.0,<3.0.0  # Email validation
orjson>=3.11.1,<4.0.0  # Fast JSON serialization
ujson>=5.11.0,<6.0.0  # Ultra fast JSON
